# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Web Umkm",
        "description": "This For API System,And You Can Try It",
        "contact": {
            "email": "muhammadalghifari4321@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/barang": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Looking Data Barang",
                "description": "Looking Data Barang",
                "operationId": "getBarang",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Insert Data Barang Admin",
                "description": "Insert Data Barang Admin",
                "operationId": "insertBarang",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_barang": {},
                                    "merk": {},
                                    "category": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/toko": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Looking Data Toko",
                "description": "Looking Data Toko",
                "operationId": "getToko",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Insert Data Toko Admin",
                "description": "Insert Data Toko Admin",
                "operationId": "insertToko",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_toko": {},
                                    "alamat": {},
                                    "user_id": {},
                                    "jenis_toko": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/toko/{id}": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Get By Id Data Toko Admin",
                "description": "Get By Id Data Toko Admin",
                "operationId": "getByIdToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Get Data And Relation",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Update Data Toko Admin",
                "description": "Update Data Toko Admin",
                "operationId": "updateToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_toko": {},
                                    "alamat": {},
                                    "user_id": {},
                                    "jenis_toko": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Delete Data Toko Admin",
                "description": "Delete Data Toko Admin",
                "operationId": "deleteToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Delete",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/count/toko": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Count Data Toko",
                "description": "Count Data Toko",
                "operationId": "countToko",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/order/toko": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Order Descending Data Toko",
                "description": "Order Descending Data Toko",
                "operationId": "orderToko",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/toko/barang": {
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Get Data Barang Toko Admin",
                "description": "Get Data Barang Toko Admin",
                "operationId": "getDataBarangByToko",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "toko": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/filter/toko": {
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Filter Data Toko By Jenis Toko",
                "description": "Filter Data Toko By Jenis Toko",
                "operationId": "filterToko",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "jenis_toko": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    ]
}